coingecko service moved into api folder
